Changes to Player and styles
added changed to the player so that it updates properly when playing a
song, also made it so that the playlist loops. Added styling and
animation to the page to make it look cleaner and stylish. Added @media
to make it mobile friendly. Replace useEffect with a function to avoid
performance errors.